refactor: codebase cleanup & root folder reorganization
- Extract media utils from App.jsx → utils/media.js (isTauri, playBlobAudio, fileToMediaUrl, playPing)
- Extract useRecording hook (3 useState + 3 useRef → hooks/useRecording.js)
- Extract useSegmentEditing hook (undo/redo, CRUD, multi-select, split/merge → hooks/useSegmentEditing.js)
- App.jsx: 2,368 → 2,057 lines (−311), useState 45 → 38, useCallback 20 → 8
- Modularize lib.rs → config.rs, bootstrap.rs, tools.rs, backend.rs, commands.rs
- Remove dead Cargo deps: tar, flate2, sha2
- Fix stray print() → logger in audio_dsp.py, mcp_server.py
- Move Docker files → deploy/, docs → docs/, scripts → scripts/
- Delete empty .codex, update STRUCTURE.md layout diagram
- Update README paths for preview.png, mcp.json
- Update docker-compose.yml build context for deploy/ location
- No version bump
