+++ docs/mcp.json
