+++ package.json
@@ -19,7 +19,10 @@
     "desktop-prod:upgrade": "bash scripts/desktop-prod.sh --keep-data",
     "build": "turbo run build",
     "start": "turbo run start",
-    "test:frontend": "node --test tests/frontend/*.test.mjs"
+    "test:frontend": "node --test tests/frontend/*.test.mjs",
+    "smoke-test": "bash scripts/smoke-test.sh",
+    "smoke-test:quick": "bash scripts/smoke-test.sh --skip-build --skip-model",
+    "smoke-test:upgrade": "bash scripts/smoke-test.sh --keep-data --skip-build"
   },
   "workspaces": [
     "frontend"
@@ -28,7 +31,7 @@
     "concurrently": "^9.2.1",
     "kill-port-process": "^4.0.2",
     "playwright": "^1.59.1",
-    "turbo": "^2.9.6",
+    "turbo": "^2.9.7",
     "typescript": "^6.0.3",
     "wait-on": "^9.0.5"
   }
